Refactor: Correction of typos
