+++ server/src/main/java/project/manager/server/controller/ContestPostController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import project.manager.server.dto.ResponseDto;
 import project.manager.server.dto.reponse.post.contest.ContestPostDto;
 import project.manager.server.dto.request.post.contest.ContestPostRequestDto;
-import project.manager.server.service.post.contast.ContestPostService;
+import project.manager.server.service.post.contest.ContestPostService;
 
 
 @RestController
