+++ server/src/main/java/project/manager/server/repository/post/contest/ContestPostRepository.java
@@ -1,11 +1,10 @@
-package project.manager.server.repository.post.contast;
+package project.manager.server.repository.post.contest;
 
 import java.time.LocalDate;
 import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
