+++ server/src/main/java/project/manager/server/service/post/contest/BenefitService.java
@@ -1,4 +1,4 @@
-package project.manager.server.service.post.contast;
+package project.manager.server.service.post.contest;
 
 import java.util.Collections;
 import java.util.List;
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import project.manager.server.domain.post.contest.Benefit;
 import project.manager.server.dto.reponse.post.contest.BenefitDto;
-import project.manager.server.repository.post.contast.BenefitRepository;
+import project.manager.server.repository.post.contest.BenefitRepository;
 
 @Service
 @Transactional
